Got rid of redundant cast warnings, reflection warnings on non-varargs call of varargs method with inexact argument type for last parameter and Class references not being typed.

+++ src/main/java/jparsec/ephem/stars/StarElement.java
@@ -244,8 +244,8 @@ public class StarElement implements Serializable
 	public static StarElement parseSimbadElement(SimbadElement s)
 	{
 		StarElement star = new StarElement(s.name, s.rightAscension, s.declination, s.parallax,
-				0, (float) (s.properMotionRA), (float) s.properMotionDEC, (float) s.properMotionRadialV,
-				Constant.J2000, EphemerisElement.FRAME.FK5);
+			0, s.properMotionRA, s.properMotionDEC, s.properMotionRadialV,
+			Constant.J2000, EphemerisElement.FRAME.FK5);
 		return star;
 	}
